+++ src/at_cmd.h
@@ -21,7 +21,17 @@
 	{                                   \
 		g_ble_uart.printf(__VA_ARGS__); \
 	}
-
 #endif
+
+#define AT_ERROR "+CME ERROR:"
+#define ATCMD_SIZE 160
+#define ATQUERY_SIZE 128
+
+#define AT_ERRNO_NOSUPP (1)
+#define AT_ERRNO_NOALLOW (2)
+#define AT_ERRNO_PARA_VAL (5)
+#define AT_ERRNO_PARA_NUM (6)
+#define AT_ERRNO_SYS (8)
+#define AT_CB_PRINT (0xFF)
 
 #endif